mdl: expand LF pinbox index space
Issue:
------
The MDL subsystem uses a process-global LF_HASH map for metadata
locks. Each THD owns one MDL_context, and the context lazily allocates
LF_PINS from the MDL map's LF_PINBOX when it first touches the map.

The old LF_PINBOX stack encoded both the free-list index and the ABA
version in a 32-bit word. The low 16 bits stored the pin array index,
and the high 16 bits stored the version. Since index 0 is reserved,
only 65535 pin indexes were usable.

A server configured with max_connections above this limit could still
fail near 65535 connection contexts that had touched MDL. This failure
is independent of the max_connections admission path: sessions that need
new MDL pins can receive ER_MDL_OUT_OF_RESOURCES even though the server
has not reached max_connections.

Solution:
---------
Extend LF_PINBOX pinstack_top_ver from 32 bits to 64 bits. The low
32 bits now store the pin array index, and the high 32 bits store the
ABA version. Keep index 0 reserved.

Raise the hard-coded LF_PINBOX capacity to the 2M level. This keeps the
implementation close to the original fixed-limit design while removing
the 16-bit index ceiling for 1M-connection configurations.

Raise the existing max_connections sysvar upper bound to 1000000. Do not
add mdl_pinbox_max_pins, MDL pinbox status variables, or startup
capacity warnings in this MR.

Docs:
Add MR-accessible Markdown documents under
Docs/features_for_opensource/21_mdl_pinbox_capacity. They capture the
source analysis, simplified design, implementation summary,
compatibility notes, reduced-limit reproduction results, verification
commands, and runtime validation SQL.

Tests:
Add gunit coverage proving the default LF pinbox can allocate past the
old 65535 usable-pin limit. Add MDL coverage proving the global MDL map
can allocate past the same old limit. Remove the MTR coverage for the
removed sysvar and status variables.

Test: Saw Mysys.LFPinboxAllocatesPastOldLimit fail before the fix.
Test: Built mysys_lf-t, mdl-t, and mysqld with Ninja.
Test: Ran Mysys.LFPinboxAllocatesPastOldLimit.
Test: Ran MDLTest.AllocatesPastOldPinboxLimit.
Test: Checked mysqld help with max-connections=1000000.
Test: Checked mdl-pinbox-max-pins is unknown with validate-config.
Test: Ran mtr 1st.
Test: Ran git diff --check.

+++ include/lf.h
@@ -77,7 +77,7 @@ struct LF_PINBOX {
   lf_pinbox_free_func *free_func;
   void *free_func_arg;
   uint free_ptr_offset;
-  std::atomic<uint32> pinstack_top_ver; /* this is a versioned pointer */
+  std::atomic<uint64> pinstack_top_ver; /* this is a versioned pointer */
   std::atomic<uint32> pins_in_array;    /* number of elements in array */
 };
